remove unused functions and variables
