+++ libpod/util.go
@@ -6,6 +6,9 @@ import (
 	"path/filepath"
 	"strings"
 	"time"
+
+	"github.com/containers/image/signature"
+	"github.com/containers/image/types"
 )
 
 // Runtime API constants
@@ -54,4 +57,22 @@ func FuncTimer(funcName string) {
 // hasTransport determines if the image string contains '://', returns bool
 func hasTransport(image string) bool {
 	return strings.Contains(image, "://")
+}
+
+// CopyStringStringMap deep copies a map[string]string and returns the result
+func CopyStringStringMap(m map[string]string) map[string]string {
+	n := map[string]string{}
+	for k, v := range m {
+		n[k] = v
+	}
+	return n
+}
+
+// GetPolicyContext creates a signature policy context for the given signature policy path
+func GetPolicyContext(path string) (*signature.PolicyContext, error) {
+	policy, err := signature.DefaultPolicy(&types.SystemContext{SignaturePolicyPath: path})
+	if err != nil {
+		return nil, err
+	}
+	return signature.NewPolicyContext(policy)
 }
